Move Oracle syscall to oracle native contract
